# task-9-10/src/infrastructure/openapi.yml
openapi: 3.0.3
info:
  title: Node.js GMP Express application
  version: 1.0.0
tags:
  - name: cart
  - name: product
paths:
  /api/profile/cart:
    post:
      tags:
        - cart
      summary: Create user cart
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '201':
          description: Returns empty cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data:
                  cart:
                    id: eb5a26af-6e4c-4f31-a9b1-3450d42ac66c
                    items: []
                  totalPrice: 0
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
    get:
      tags:
        - cart
      summary: Get user cart or create if it is missing
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Returns user cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data:
                  cart:
                    id: eb5a26af-6e4c-4f31-a9b1-3450d42ac66c
                    items:
                      - product:
                          id: 891389f0-4312-42d6-a650-6fda0959c734
                          title: Book
                          description: Interesting book
                          price: 200
                        count: 2
                  totalPrice: 400
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
    put:
      tags:
        - cart
      summary: Update user cart
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cart'
      responses:
        '200':
          description: Cart can be updated in the following ways - 1) products can be added 2) products can be dropped 3) amount of product might have changed. Request body to be provided contains a snapshot of cart after changes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data:
                  cart:
                    id: eb5a26af-6e4c-4f31-a9b1-3450d42ac66c
                    items:
                      - product:
                          id: 891389f0-4312-42d6-a650-6fda0959c734
                          title: Book
                          description: Interesting book
                          price: 200
                        count: 2
                  totalPrice: 400
                error: null
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Product name should be string, but not number
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
    delete:
      tags:
        - cart
      summary: Empty user cart
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Returns success = true if cart was successfully emptied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessResponse'
              example:
                data:
                  success: true
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
  /api/profile/cart/checkout:
    post:
      tags:
        - cart
      summary: Create an order
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
              example:
                data:
                  order:
                    id: 6c36d6fa-f694-4f9c-9b2c-6f7049d38f4a
                    userId: 6dc52b3c-de7e-431a-84b8-0ec56e0774d4
                    cartId: cadff0c5-0079-4db8-b6bf-84c9c2633ca3
                    items:
                      - product:
                          id: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
                          title: Book
                          description: Interesting book
                          price: 200
                        count: 2
                      - product:
                          id: afdd68c4-d359-45e6-b9fd-c8fdb2a162a0
                          title: Pen
                          description: Cute pen
                          price: 20
                        count: 5
                    payment:
                      type: paypal
                      address: London
                      creditCard: 1234-1234-1234-1234
                    delivery:
                      type: post
                      address: London
                    comments: ''
                    status: created
                    totalPrice: 500
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
  /api/products:
    get:
      tags:
        - product
      summary: Returns a list of products
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Returns a list of all products available
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductsResponse'
              example:
                data:
                  - id: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
                    title: Book
                    description: Interesting book
                    price: 200
                  - id: afdd68c4-d359-45e6-b9fd-c8fdb2a162a0
                    title: Pen
                    description: Cute pen
                    price: 20
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
  /api/products/{productId}:
    get:
      tags:
        - product
      summary: Returns single product
      parameters:
        - name: productId
          in: path
          description: Id (uuid) of product to return
          required: true
          schema:
            type: number
        - in: header
          name: x-user-id
          schema:
            type: string
            format: uuid
          required: true
          description: User id (uuid)
          example: 'eb5a26af-6e4c-4f31-a9b1-3450d42ac66c'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
              example:
                data:
                  id: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
                  title: Book
                  description: Interesting book
                  price: 200
                error: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: Header x-user-id is missing or no user with such id
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: No product with such id
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
              example:
                data: null
                error:
                  message: 'Ooops, something went wrong'
components:
  schemas:
    Cart:
      type: object
      properties:
        id:
          type: string
          description: Cart id (uuid)
          example: dd5ec5ab-deaa-419c-8a6b-7e67b1f7ec87
        items:
          type: array
          description: Items added to cart
          items:
            $ref: '#/components/schemas/CartItem'
    CartItem:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'
        count:
          type: number
          description: Total count of specific products
          example: 2
    Product:
      type: object
      properties:
        id:
          type: string
          description: Product id (uuid)
          example: 5c293ad0-19d0-41ee-baa3-4c648f9f7697
        title:
          type: string
          description: Product name
          example: Book
        description:
          type: string
          description: Product description
          example: Interesting book
        price:
          type: number
          description: Product price
          example: 200
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        cartId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        payment:
          type: object
          properties:
            type:
              type: string
            address:
              type: string
            creditCard:
              type: string
        delivery:
          type: object
          properties:
            type:
              type: string
            address:
              type: string
        comments:
          type: string
        status:
          type: string
        totalPrice:
          type: number
    CartResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            cart:
              $ref: '#/components/schemas/Cart'
            totalPrice:
              type: number
        error:
          $ref: '#/components/schemas/ErrorResponse'
    CheckoutResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            order:
              $ref: '#/components/schemas/Order'
        error:
          $ref: '#/components/schemas/ErrorResponse'
    ProductResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Product'
        error:
          $ref: '#/components/schemas/ErrorResponse'
    ProductsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        error:
          $ref: '#/components/schemas/ErrorResponse'
    EmptySuccessResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            success:
              type: boolean
        error:
          $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      nullable: true
      properties:
        message:
          type: string